% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemarriageR.R
\name{check_valid_groups}
\alias{check_valid_groups}
\title{Check whether a pair of groups is valid}
\usage{
check_valid_groups(group1, group2)
}
\description{
Raises an error if the inputs are not a valid pair of groups.
}
