% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemarriageR.R
\name{check_valid_preferences}
\alias{check_valid_preferences}
\title{Check whether a pair of preference tables is valid}
\usage{
check_valid_preferences(group1, group2, pref1, pref2)
}
\description{
Checks whether pref1 and pref2 are data frames that correspond to each other 
and groups 1 & 2 in terms of size and content, raises an error otherwise.
}
