% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemarriageR.R
\name{randomise_preferences}
\alias{randomise_preferences}
\title{Generate a random pair of preference tables}
\usage{
randomise_preferences(group1, group2)
}
\arguments{
\item{group1}{character vector of distinct nonempty strings}

\item{group2}{character vector of distinct nonempty strings}
}
\value{
Returns a list containing the two preference tables .
}
\description{
Produces a random pair of preference tables for group1 and group2
}
