% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemarriageR.R
\name{read_pref}
\alias{read_pref}
\title{Read in a preference table}
\usage{
read_pref(file)
}
\arguments{
\item{file}{the name of the file containing the preference table}
}
\value{
A \code{data.frame} containing the data in the file
}
\description{
This is a convenience wrapper for \code{read.csv} with \code{stringsAsFactors = FALSE} .
}
