% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemarriageR.R
\name{show}
\alias{show}
\title{Displays instances of S4 objects of type \code{\link{stablemarriageclass}}}
\usage{
\S4method{show}{stablemarriageclass}(object)
}
\arguments{
\item{object}{An instance of an S4 object of type \code{\link{stablemarriageclass}}.}
}
\description{
Displays the groups and their preference tables, as well as a
  matching if found
}
\seealso{
\code{\link{stablemarriageclass}}
}
