% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemarriageR.R
\name{stablemarriage}
\alias{stablemarriage}
\title{Solves instances of the stable marriage problem}
\usage{
stablemarriage(
  group1,
  group2,
  pref1 = NULL,
  pref2 = NULL,
  validate_inputs = TRUE
)
}
\arguments{
\item{group1}{character vector of distinct nonempty strings.}

\item{group2}{character vector of distinct nonempty strings.}

\item{pref1}{data frame with a column for each member of \code{group1} 
containing the members of \code{group2} sorted from most to least preferred. 
Optional - if not supplied preference tables will be randomised.}

\item{pref2}{data frame with a column for each member of \code{group2} 
containing the members of \code{group1} sorted from most to least preferred.
Optional - if not supplied preference tables will be randomised.}

\item{validate_inputs}{logical value (default: \code{TRUE}) indicating whether
to check that the groups and preference tables supplied are valid.}
}
\value{
An instance of an S4 object of type \code{\link{stablemarriageclass}}, 
  containing a stable matching for the problem.
}
\description{
Solves instances of the stable marriage problem as described in: 
  'Stable Marriage and Its Relation to Other Combinatorial Problems : 
  An Introduction to the Mathematical Analysis of Algorithms' 
  by Donald E. Knuth and Martin Goldstein.
}
\seealso{
\code{\link{stablemarriageclass}}, \code{\link{randomise_preferences}}
}
