% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablemarriageR.R
\docType{class}
\name{stablemarriageclass}
\alias{stablemarriageclass}
\title{An S4 class to represent an instance of the stable marriage problem}
\description{

}
\section{Slots}{

\describe{
\item{\code{group1}}{character vector of distinct nonempty strings.}

\item{\code{group2}}{character vector of distinct nonempty strings.}

\item{\code{pref1}}{data frame with a column for each member of \code{group1} 
containing the members of \code{group2} sorted from most to least preferred.}

\item{\code{pref2}}{data frame with a column for each member of \code{group2} 
containing the members of \code{group1} sorted from most to least preferred.}

\item{\code{matching}}{data frame with members of \code{group2} as column names containing 
a bijection between \code{group2} and \code{group1}.}
}}

\seealso{
\code{\link{stablemarriage}}
}
