\name{stablemarriageR}
\alias{stablemarriageR}
\docType{package}
\title{
  Solves the stable marriage problem.
}
\description{
Package implementing the fundamental algorithm to solve the stable marriage problem, as described in 'Stable Marriage and Its Relation to Other Combinatorial Problems : An Introduction to the Mathematical Analysis of Algorithms' by Donald E. Knuth and Martin Goldstein.
}
\author{
Connie Trojan

Maintainer: Connie Trojan <c.trojan1@lancaster.ac.uk>
}
\keyword{ package }
\seealso{
  \code{\link{stablemarriage}}
}